fix: move webgpu_limits.json into src for CRA import scope
CRA/webpack ModuleScopePlugin blocks imports outside src/. The shared
WebGPU limits contract JSON now lives at src/contracts/webgpu_limits.json
with updated imports, verify script, and docs.
