+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -6,7 +6,7 @@
 
 - [ ] `npm test -- --watchAll=false --ci` passes locally
 - [ ] If **C++ or `wasm_renderer/bridge/`** changed: `npm run wasm:build` then **`npm run wasm:validate`**
-- [ ] If **device limits / bind group** changed: update `contracts/webgpu_limits.json` + `wasm_renderer/device.cpp`; run `npm run verify:device-policy`
+- [ ] If **device limits / bind group** changed: update `src/contracts/webgpu_limits.json` + `wasm_renderer/device.cpp`; run `npm run verify:device-policy`
 - [ ] If **WGSL shaders** changed: `python3 scripts/wgsl_precommit_gate.py --files <paths>`
 
 ## WASM / build notes
