+++ memory/2026-07-19.md
@@ -42,7 +42,7 @@
 - WASM promotion gates 1–4 evidence collection
 
 ## Build & tooling hardening (#965)
-- contracts/webgpu_limits.json + webgpuLimitsContract.test.ts
+- src/contracts/webgpu_limits.json + webgpuLimitsContract.test.ts (moved from contracts/ for CRA import scope)
 - INITIAL_MEMORY rejected; ASYNCIFY ~40KB wasm cost documented
 - docs/TOOLCHAIN_DECISION.md (stay CRA)
 - PR template wasm:validate checklist
